Refactor console greeting and improve font size defaults
- Extract console greeting logic to a separate utility function
- Modify font size selection to use device-specific defaults
- Update HNLiveTerminal to use new console greeting utility
- Remove hardcoded console greeting in main component
- Add intelligent font size selection for PWA and mobile devices
